move imgui impl of painter to `PainterImGui.cpp`
